requirements.txt
numpy
Pillow
dropbox
